requirements.txt
Django>=3.2,<4.0
djangorestframework
celery
redis
django-environ
// ...existing code...
